+++ gulpfile.js
@@ -22,3 +22,7 @@ gulp.task(config.build.images.command, getTask("build-images"));
 gulp.task(config.build.fonts.command, getTask("build-fonts"));
 gulp.task("delete:dist", getTask("delete-dist"));
 gulp.task("build:dist", getTask("build-dist"));
+
+exports.run = function (command) {
+    gulp.start(command);
+};